Rewrite dehyphenate's parser.
It is a lot less brittle now, and hopefully easier to read.

Additionally:

The way page breaks are handled has changed: previously, they were
translated to a line break followed by a form feed; now they are
translated to just a form feed. This is maybe more correct?

A new output format has been added: 'orig'. This converts form feeds to
double-newlines, and soft hyphens to hyphens--hopefully round-tripping
back to the original encoding.